Lib

Documentation

newtype a :=> b

a is the Cxt of b

Constructors

Cxt

Fields

runCxt :: a -> b

data V v a

Variables of type a with ids of type v

Constructors

V

Fields

varId :: v

varVal :: a

newtype Pattern v a b

Pattern bindings where variables have id type v and the pattern is bound from matching on a's to a result of type b.

Constructors

Pattern

Fields

runPattern :: forall c d. (a -> Either c b) :=> (V v a -> (V v b :=> PExpr v d) -> Either c (PExpr v d))

idPatt :: Pattern v a a

Identity pattern

data PExpr v a where

Expressions with variables, lambda/let binding, and literals

Also with Patterns, but not used much

Constructors

PLit :: a -> PExpr v a

PVar :: ((V v a :=> PExpr v a) :~: t) -> PExpr v t

PLambda :: ((a -> b) :~: t) -> v -> Pattern v a a -> PExpr v (V v a :=> PExpr v b) -> PExpr v t

PLet :: v -> Pattern v a a -> PExpr v (V v a :=> PExpr v b) -> PExpr v a -> PExpr v b

PToCxt :: ((a :=> b) :~: t) -> PExpr v (a -> b) -> PExpr v t

data Expr v a where

Expressions with variables, lambda/let binding, and literals

Constructors

Lit :: a -> Expr v a

Var :: ((V v a :=> Expr v a) :~: t) -> Expr v t

Lambda :: ((a -> b) :~: t) -> v -> Expr v (V v a :=> Expr v b) -> Expr v t

Let :: v -> Expr v (V v a :=> Expr v b) -> Expr v a -> Expr v b

ToCxt :: ((a :=> b) :~: t) -> Expr v (a -> b) -> Expr v t

App :: Expr v (a -> b) -> Expr v a -> Expr v b

evalp :: PExpr v a -> a

Evaluate a PExpr

 λ> eval (PLambda Refl () idPatt (PVar Refl)) (10 :: Int)
 10

eval :: Expr v a -> a

Evaluate an Expr

 λ> eval (Lambda Refl () (Var Refl)) (10 :: Int)
 10

fromRightVoid :: Either Void a -> a

Lazy unpacking of Right